- Added support for spells that target "all" enemies/allies - Added support for spells that apply different effect levels per spell level - Starting hero ids can now be specified via name (String) as opposed to ID (int) - Added limited elemental affinities support - Added support for effects to prevent turns - Cinematic creator now has a remember mouse location feature (via middle click?) - Map editor now displays referred to object name as opposed to ID (Usability) - Planner now displays affinities, weapon types and condtions - The "tree" in the planner now allows for duplicating an object - Planner now allows for bookmarks to be saved when locating files (Usability) - Added new Planner tab that allows non-map-location condtions to be specified to drive a "trigger" - Other stuff

+++ CommRPG/src/mb/fc/game/Camera.java
@@ -31,14 +31,14 @@ public class Camera
 		return viewport;
 	}
 
-	public int getLocationX()
+	public float getLocationX()
 	{
-		return (int) viewport.getX();
+		return viewport.getX();
 	}
 
-	public int getLocationY()
+	public float getLocationY()
 	{
-		return (int) viewport.getY();
+		return viewport.getY();
 	}
 
 	public void setLocation(float x, float y)
